+++ src/App.js
@@ -1,43 +1,15 @@
-import React,{useCallback} from 'react';
-import ReactFlow,{useEdgesState,useNodesState,addEdge,MiniMap,
-  Controls,
-  Background,} from 'reactflow';
- 
-import 'reactflow/dist/style.css';
- 
-const initialNodes = [
-  { id: '1', position: { x: 0, y: 0 }, data: { label: '1' } },
-  { id: '2', position: { x: 0, y: 100 }, data: { label: '2' } },
-  { id: '3', position: { x: 0, y: 250 }, data: { label: '3' } },
-];
-const initialEdges = [{ id: 'e1-2', source: '1', target: '2' },
-{ id: 'e2-3', source: '2', target: '3' },];
+import React from 'react';
+import { BasicIntro } from './components/BasicIntro';
+
+
  
 export default function App() {
 
-  const [nodes, setNodes, onNodesChange] = useNodesState(initialNodes);
-  const [edges, setEdges, onEdgesChange] = useEdgesState(initialEdges);
 
-  const onConnect = useCallback(
-    (params) => setEdges((eds) => addEdge(params, eds)),
-    [setEdges],
-  );
 
   return (
-    <div style={{ width: '100vw', height: '100vh' }}>
-      {/* <ReactFlow nodes={initialNodes} edges={initialEdges} /> */}
-
-      <ReactFlow
-        nodes={nodes}
-        edges={edges}
-        onNodesChange={onNodesChange}
-        onEdgesChange={onEdgesChange}
-        onConnect={onConnect}
-      >
-       <Controls />
-        <MiniMap />
-        <Background variant="dots" gap={12} size={1} />
-        </ReactFlow>
-    </div>
+   <div>
+    <BasicIntro />
+   </div>
   );
 }
